+++ lib/minitest/mock_expectations/assertions.rb
@@ -98,10 +98,33 @@ module Minitest
     #     @post.add_comment("Thanks for sharing this.")
     #     @post.add_comment("Thanks!")
     #   end
+    #
+    #   assert_called_with(@post, :add_comment, [["Thanks for sharing this."]]) do
+    #     @post.add_comment(["Thanks for sharing this."])
+    #   end
+    #
+    #   assert_called_with(@post, :add_comment, [["Thanks for sharing this.", "Thanks!"]]) do
+    #     @post.add_comment(["Thanks for sharing this.", "Thanks!"])
+    #   end
+    #
+    #   assert_called_with(@post, :add_comment, [["Thanks for sharing this."], {body: "Thanks!"}]) do
+    #     @post.add_comment(["Thanks for sharing this."], {body: "Thanks!"})
+    #   end
+    #
+    #   assert_called_with(@post, :add_comment, [[["Thanks for sharing this."]], [{body: "Thanks!"}]]) do
+    #     @post.add_comment(["Thanks for sharing this."])
+    #     @post.add_comment({body: "Thanks!"})
+    #   end
+    #
+    #   assert_called_with(@post, :add_comment, [[["Thanks for sharing this."]], [["Thanks!"]]]) do
+    #     @post.add_comment(["Thanks for sharing this."])
+    #     @post.add_comment(["Thanks!"])
+    #   end
+    #
     def assert_called_with(object, method_name, arguments, returns: nil)
       mock = Minitest::Mock.new
 
-      if arguments.all? { |argument| argument.is_a?(Array) }
+      if arguments.size > 1 && arguments.all? { |argument| argument.is_a?(Array) }
         arguments.each { |argument| mock.expect(:call, returns, argument) }
       else
         mock.expect(:call, returns, arguments)